ssm/generator/requirements.txt
httpx>=0.27
PyMuPDF>=1.24
python-dotenv>=1.0
flask>=3.0
